:recycle: Move error content to a component
